Create CameraFocusPoint and use in CentralHub.
This new scene allows for panning the camera to a custom position when
the player enters a trigger area. The main idea is to use it to pan the
camera over to a fixed central location for the warden fight. Another
idea is to use it before the fight to draw attention to the central lock
as the player moves around the hub.