Create basic intro fall sequence for warden.
So far this is mostly sprite animations, with an initial fall (where the
warden crashes through the ceiling in a crouched position), the landing
(some rumble/screenshake on a timer), and the uncrouch (a multi-tiered
animation with two uncrouch phases, as well as a phase where the
lights/eyes turn on).

Future commits will add extra niceties: dust puff sprites, debris
particles, sounds, etc.