Organize rooms in World node by section/sector.
Rather than having all rooms laid out directly underneath World, they
are now placed underneath intermediate nodes that are themselves
children of World. This extra level of indirection allows for querying
which sector a specific room is in without having to resort to things
like node name prefixes (which might not always be there) or node groups
(which I might forget to add when adding new rooms).

This may be used in future commits to determine when to phase
sector-specific music in and out (or at least to determine which
sector-specific layer to play on startup). This may also be used in the
future to enable certain optimizations (e.g. removing inaccessible
sectors/sectors from the scene tree).