Adjust pillar position in SectorTwo_14.
This pillar is supposed to teach the player to wall jump along a single
wall, but the previous position of this pillar allowed for the player to
wall jump off the opposite wall and dash over it. The new position is
far enough to the left to prevent this and force the player to jump
along the pillar wall.