Raise collapsable platform in SectorOne_18.
This allows for adding spikes directly underneath the platform without
having tilemap overlap issues.

TECHNICALLY it seems as though it might be possible for a player with
only the wall jump to sus their way through this gap, but it's very hard
if so, and the player needs the dash in the next room anyway so it's fine.