Add glitch effect to enemies spawning into arenas.
Rather than enemies appearing (with their hitboxes) instantly, they now
have dedicated spawn states whereby their presence is announced via a
nice little glitch effect. This should hopefully better alert the player
to their presence and give the player an opportunity to get out of the
way if an enemy happens to be spawning in on top of them.

This fixes #31