Fix memory leaks.
There were three main leaks:

1) The GameInterface instance exported from TitleScreen, fixed by
   replacing the exported variable with a hardcoded preload.

2) The arena enemies, fixed by manually freeing all the enemies in the
   phase data array on tree exit.

3) ShaderManager, fixed by creating a dedicated node and explicitly
   adding it to the tree for any node that was using it.

Note that for 2), we can't avoid the fact that orphan nodes will be
present throughout normal gameplay, since we are explicitly removing
arena enemies from the tree so that we can save them in an array to
spawn them in later. This is fine, so long as we clean them up before
scene changes, which we do now. This prevents the actual leak when going
to and from the main menu and actual game.

This fixes #30