Ground player during warden intro cutscene.
Rather than allow the player to keep moving through the air, simply
force them to the ground once they pass the trigger point. This avoids
the corner case where the player could end up right beneath the warden
as it comes crashing down (and is more in line with how most games seem
to handle this kind of thing).